dinedash-backend/requirements.txt
so we can work Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
dj-rest-auth==6.0.0
django-allauth==65.0.2
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
boto3==1.35.24
django-storages==1.14.4
Pillow==11.1.0